Fix some issue I had with karma
