+++ karma.conf.ts
@@ -1,0 +1,73 @@
+/* eslint-disable @typescript-eslint/no-unsafe-member-access */
+import { Config, ConfigOptions } from 'karma';
+import path from 'path';
+import webpack from 'webpack';
+
+import { WEBPACK_CONFIG } from './webpack.config';
+
+/**
+ * Small webpack plugin to remove typescript declaration files (.d.ts) from the
+ * list of assets.
+ *
+ * karma-webpack is using this list to know which file to load for tests, which
+ * contains both the path to test files compiled to javascript, and the path to
+ * typescript declaration files. Webpack output path is set to a temporary
+ * directory, and the javascript files are inside this directory. The typescript
+ * declaration files are not in this temporary directory (they are emitted in
+ * `chemiscope/dist/` by tsc), and for a still unknown reason that can cause
+ * karma not to found these file in some very specific cases (macOS and/or ZSH
+ * seems to be required to hit this issue).
+ *
+ * Since we do not care about the `.d.ts` files in karma, this plugin works
+ * around the issue by removing said files from the assets list. This can break
+ * if webpack plugin API changes, or if karma-webpack stops using the assets.
+ */
+class RemoveDeclarationsFromAssets implements webpack.WebpackPluginInstance {
+    apply(compiler: webpack.Compiler) {
+        compiler.hooks.done.tap('RemoveDeclarationsFromAssets', (stats) => {
+            const toRemove = [];
+            for (const path in stats.compilation.assets) {
+                if (path.endsWith('.d.ts')) {
+                    toRemove.push(path);
+                }
+            }
+
+            for (const key of toRemove) {
+                delete stats.compilation.assets[key];
+            }
+        });
+    }
+}
+
+WEBPACK_CONFIG.plugins?.push(new RemoveDeclarationsFromAssets());
+
+module.exports = (config: Config) => {
+    config.set({
+        files: [path.resolve(__dirname, 'tests/**/*.test.ts')],
+        frameworks: ['webpack', 'mocha', 'detectBrowsers'],
+
+        preprocessors: {
+            'tests/**/*.test.ts': 'webpack',
+        },
+        reporters: ['progress'],
+        singleRun: true,
+
+        webpack: WEBPACK_CONFIG,
+        webpackMiddleware: {
+            stats: 'errors-only',
+        },
+
+        detectBrowsers: {
+            postDetection: function (availableBrowsers: string[]) {
+                // Remove IE
+                const IEindex = availableBrowsers.indexOf('IE');
+                if (IEindex !== -1) {
+                    availableBrowsers.splice(IEindex);
+                }
+                return availableBrowsers;
+            },
+            preferHeadless: false,
+            usePhantomJS: false,
+        },
+    } as ConfigOptions);
+};
