+++ package.json
@@ -18,7 +18,7 @@
   },
   "scripts": {
     "test": "npm run test:karma && npm run test:lint",
-    "test:karma": "karma start tests/karma.conf.ts",
+    "test:karma": "karma start",
     "test:lint": "eslint src && prettier --check src",
     "build": "rimraf dist && webpack --mode production && npm run merge-dts",
     "download-examples": "ts-node ./utils/download-examples.ts",
@@ -36,7 +36,7 @@
     "@types/jquery": "^3.5.6",
     "@types/markdown-it": "^12.0.3",
     "@types/mocha": "^8.2.3",
-    "@types/node": "^15.14.1",
+    "@types/node": "^16.3.0",
     "@types/pako": "^1.0.2",
     "@types/plotly.js": "^1.54.11",
     "@types/tmp": "^0.2.1",
@@ -62,9 +62,9 @@
     "karma-firefox-launcher": "^2.1.1",
     "karma-mocha": "^2.0.1",
     "karma-safari-launcher": "^1.0.0",
-    "karma-webpack": "^5.0.0",
+    "karma-webpack": "git+https://github.com/ryanclark/karma-webpack#ef7edb6b6756fb563871eaff88ca876892694896",
     "markdown-it": "^12.1.0",
-    "mocha": "^8.1.2",
+    "mocha": "^9.0.0",
     "node-loader": "^2.0",
     "plotly.js": "1.58.4",
     "popper.js": "^1.16.1",
@@ -74,11 +74,11 @@
     "style-loader": "^2.0.0",
     "tmp": "^0.2.1",
     "ts-loader": "^9.2.3",
-    "ts-node": "^9.0.0",
+    "ts-node": "^10.0.0",
     "typedoc": "^0.19",
     "typedoc-plugin-external-module-name": "^4",
     "typescript": "^4.3.5",
-    "webpack": "^5.43.0",
+    "webpack": "^5.44.0",
     "webpack-cli": "^4.7.2",
     "webpack-dev-server": "4.0.0-beta.3"
   },
